Squash compilation warning messages

+++ source/libcheetah/src/saveCXI.cpp
@@ -12,6 +12,7 @@
 #include <vector>
 
 #include "saveCXI.h"
+#include "util.h"
 
 herr_t
 cheetahHDF5ErrorHandler(hid_t, void *)
@@ -63,7 +64,7 @@ static T * generateThumbnail(const T * src,const int srcWidth, const int srcHeig
 	  res += src[yy*srcWidth+xx];
 	} 
       }
-      dst[y*dstWidth+x] = res/(scale*scale);
+      dst[y*dstWidth+x] = (short int) res/(scale*scale);
     }
   }
   return dst;
